requirements.txt
Flask
notebook
mysql-connector-python
sqlalchemy

